pyproject.toml
[project]
name = "aqa"
version = "0.1.0"
description = "Lightweight site explorer + test generator"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "requests",
  "beautifulsoup4",
  "PyYAML",
  "pytest"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
